Support half-star ratings on brews and visits
StarRating is now a custom widget with two click zones per star (left
half = 0.5, right half = 1.0). Brew.Rating and CoffeeShopVisit.Rating
move from int to decimal (existing int values deserialize cleanly).
Read-only call sites pass averages directly so the widget snaps to the
nearest half visually.

+++ Models/CoffeeShopVisit.cs
@@ -35,7 +35,8 @@ public class CoffeeShopVisit
     public string? CoffeeOrigin { get; set; }
 
     public string? Notes { get; set; }
-    public int Rating { get; set; }
+    /// <summary>Note de 0 à 5 ⭐, par pas de 0,5 (demi-étoiles autorisées).</summary>
+    public decimal Rating { get; set; }
     public decimal? Price { get; set; }
     public string? Currency { get; set; } = "EUR";
     public string? PhotoDataUrl { get; set; }
